+++ src/extension_commands.ml
@@ -88,6 +88,7 @@ let _upgrade_ocaml_lsp_server =
       match
         Ocaml_lsp.is_version_up_to_date
           (Extension_instance.ocaml_lsp instance |> Option.value_exn)
+          (Extension_instance.sandbox instance)
           (Extension_instance.ocaml_version_exn instance)
       with
       | Ok () -> Promise.return ()
@@ -619,6 +620,7 @@ end = struct
     match
       Ocaml_lsp.is_version_up_to_date
         ocaml_lsp
+        (Extension_instance.sandbox instance)
         (Extension_instance.ocaml_version_exn instance)
     with
     | Ok () -> ()
@@ -845,6 +847,7 @@ module Copy_type_under_cursor = struct
     match
       Ocaml_lsp.is_version_up_to_date
         ocaml_lsp
+        (Extension_instance.sandbox instance)
         (Extension_instance.ocaml_version_exn instance)
     with
     | Ok () -> ()