+++ src/ocaml_lsp.mli
@@ -3,7 +3,13 @@ open Import
 type t
 
 val of_initialize_result : LanguageClient.InitializeResult.t -> t
-val is_version_up_to_date : t -> Ocaml_version.t -> (unit, [ `Msg of string ]) result
+
+val is_version_up_to_date
+  :  t
+  -> Sandbox.t
+  -> Ocaml_version.t
+  -> (unit, [ `Msg of string ]) result
+
 val can_handle_switch_impl_intf : t -> bool
 val can_handle_infer_intf : t -> bool
 val can_handle_typed_holes : t -> bool
@@ -28,20 +34,39 @@ module OcamllspSettingCodeLens : sig
   val create : ?forNestedBindings:bool -> enable:bool -> unit -> t
 end
 
+module OcamllspSettingInlayHints : sig
+  include Ojs.T
+
+  val hintPatternVariables : t -> bool option
+  val hintLetBindings : t -> bool option
+  val hintFunctionParams : t -> bool option
+
+  val create
+    :  ?hintPatternVariables:bool
+    -> ?hintLetBindings:bool
+    -> ?hintFunctionParams:bool
+    -> unit
+    -> t
+end
+
 module OcamllspSettings : sig
   include Ojs.T
 
-  val codelens : t -> OcamllspSettingCodeLens.t option
   val extendedHover : t -> OcamllspSettingEnable.t option
   val standardHover : t -> OcamllspSettingEnable.t option
+  val codelens : t -> OcamllspSettingCodeLens.t option
   val duneDiagnostics : t -> OcamllspSettingEnable.t option
+  val inlayHints : t -> OcamllspSettingInlayHints.t option
   val syntaxDocumentation : t -> OcamllspSettingEnable.t option
+  val shortenMerlinDiagnostics : t -> OcamllspSettingEnable.t option
 
   val create
-    :  codelens:OcamllspSettingCodeLens.t option
-    -> extendedHover:OcamllspSettingEnable.t option
+    :  extendedHover:OcamllspSettingEnable.t option
     -> standardHover:OcamllspSettingEnable.t option
+    -> codelens:OcamllspSettingCodeLens.t option
     -> duneDiagnostics:OcamllspSettingEnable.t option
+    -> inlayHints:OcamllspSettingInlayHints.t option
     -> syntaxDocumentation:OcamllspSettingEnable.t option
+    -> shortenMerlinDiagnostics:OcamllspSettingEnable.t option
     -> t
 end