+++ src/treeview_switches.mli
@@ -1,2 +1,6 @@
 (** Register the ocaml-switches tree view *)
-val register : Vscode.ExtensionContext.t -> Extension_instance.t -> unit
+val register
+  :  Vscode.ExtensionContext.t
+  -> Extension_instance.t
+  -> assets:Extension_assets.t
+  -> unit